Refactor fix types
